--- Editor/data/scene.js
@@ -1,128 +1,0 @@
-export const arpObjects = {
-    "happy-sun": {
-        "image": "./images/happy/happy-sun.png",
-        "mood": "happy",
-        "chordName": "Cmajor9",
-        "synthName" : "polySynth"
-    },
-    "happy-cloud": {
-        "image": "./images/happy/happy-cloud.png",
-        "mood": "happy",
-        "chordName": "Gmajor",
-        "synthName" : "polySynth"
-    },
-    "happy-tree": {
-        "image": "./images/happy/happy-tree.png",
-        "mood": "happy",
-        "chordName": "Cmajor",
-        "synthName" : "polySynth"
-    },
-    "happy-rock": {
-        "image": "./images/happy/happy-rock.png",
-        "mood": "happy",
-        "chordName": "C5",
-        "synthName" : "polySynth"
-    },
-    "sad-cloud": {
-        "image": "./images/sad/sad-cloud.png",
-        "mood": "sad",
-        "chordName": "COctave",
-        "synthName" : "polySynth"
-    },
-    "sad-cloud-02": {
-        "image": "./images/sad/sad-cloud-02.png",
-        "mood": "sad",
-        "chordName": "C5",
-        "synthName" : "polySynth"
-    },
-    "sad-birds": {
-        "image": "./images/sad/sad-birds.png",
-        "mood": "sad",
-        "chordName": "Cminor",
-        "synthName" : "polySynth"
-    },
-    "sad-lightning": {
-        "image": "./images/sad/sad-lightning.png",
-        "mood": "sad",
-        "chordName": "Cminor9",
-        "synthName" : "polySynth"
-    },
-    "angry-man": {
-        "image": "./images/angry/angry-man.png",
-        "mood": "angry",
-        "chordName": "COctave",
-        "synthName" : "polySynth"
-    },
-    "angry-face": {
-        "image": "./images/angry/angry-face.png",
-        "mood": "angry",
-        "chordName": "C5",
-        "synthName" : "polySynth"
-    },
-    "angry-swear": {
-        "image": "./images/angry/angry-swear.png",
-        "mood": "angry",
-        "chordName": "CminorMajor7",
-        "synthName" : "polySynth"
-    },
-    "nostalgic-kid": {
-        "image": "./images/nostalgic/nostalgic-kid.png",
-        "mood": "nostalgic",
-        "chordName": "Cmajor7",
-        "synthName" : "polySynth"
-    },
-    "nostalgic-swing": {
-        "image": "./images/nostalgic/nostalgic-swing.png",
-        "mood": "nostalgic",
-        "chordName": "Csus2",
-        "synthName" : "polySynth"
-    },
-    "nostalgic-tree": {
-        "image": "./images/nostalgic/nostalgic-tree.png",
-        "mood": "nostalgic",
-        "chordName": "C5",
-        "synthName" : "polySynth"
-    },
-    "spooky-baphomet": {
-        "image": "./images/spooky/spooky-baphomet.png",
-        "mood": "spooky",
-        "chordName": "COctave",
-        "synthName" : "polySynth"
-    },
-    "spooky-skull": {
-        "image": "./images/spooky/spooky-skull.png",
-        "mood": "spooky",
-        "chordName": "CWholetoneArp",
-        "synthName" : "polySynth"
-    },
-    "spooky-ouroboros": {
-        "image": "./images/spooky/spooky-ouroboros.png",
-        "mood": "spooky",
-        "chordName": "CWholetoneAscending",
-        "synthName" : "polySynth"
-    },
-    "serene-crystal": {
-        "image": "./images/serene/serene-crystal.png",
-        "mood": "serene",
-        "chordName": "C7sus4",
-        "synthName" : "polySynth"
-    },
-    "serene-lotus": {
-        "image": "./images/serene/serene-lotus.png",
-        "mood": "serene",
-        "chordName": "Csus4",
-        "synthName" : "polySynth"
-    },
-    "serene-buddha": {
-        "image": "./images/serene/serene-buddha.png",
-        "mood": "serene",
-        "chordName": "F5",
-        "synthName" : "polySynth"
-    },
-    "serene-om": {
-        "image": "./images/serene/serene-om.png",
-        "mood": "serene",
-        "chordName": "COctave",
-        "synthName" : "polySynth"
-    }
-};
